TUITextView now responds to -editable property
TUITextView ignored -editable before, and -userInteractionEnabled would
completely shut off dictionary and highlighting functions.  By adding
check for edit-ability to the _isKey function (and a property to
TUITextEditor), the text view must pass two tests before it can become
a first responder and show the cursor to accepts key events.  It is a
bit fake-ish though, because clearly the renderer must respond to click
events, otherwise double clicks on words don't go through correctly,
but at least it guts the ability of the user to add text to a text view
they shouldn't be adding text to.  It's possible that you may not want
to expose the editable property I've defined in case it accidentally
disables some poor chap's text view, but it's worth a look.